fix: progress.

+++ dtos/k8sServiceSettingsDto.go
@@ -1,11 +1,11 @@
 package dtos
 
 type K8sServiceSettingsDto struct {
-	LimitCpuCores      int    `json:"limitCpuCores" validate:"required"`
-	LimitMemoryMB      int    `json:"limitMemoryMB" validate:"required"`
-	EphemeralStorageMB int    `json:"ephemeralStorageMB" validate:"required"`
-	ReplicaCount       int    `json:"replicaCount" validate:"required"`
-	DeploymentStrategy string `json:"deploymentStrategy" validate:"required"` // "rolling", "recreate"
+	LimitCpuCores      float32 `json:"limitCpuCores" validate:"required"`
+	LimitMemoryMB      int     `json:"limitMemoryMB" validate:"required"`
+	EphemeralStorageMB int     `json:"ephemeralStorageMB" validate:"required"`
+	ReplicaCount       int32   `json:"replicaCount" validate:"required"`
+	DeploymentStrategy string  `json:"deploymentStrategy" validate:"required"` // "rolling", "recreate"
 }
 
 func K8sServiceSettingsDtoExampleData() K8sServiceSettingsDto {
